Split pages into separate files, fix responsiveness, add rest authentication function, customise selection, add linting, add enter key submission, change default font to Jetbrains, prevent Ctrl+S
