+++ frontend/src/pages/register/Register.jsx
@@ -2,6 +2,7 @@ import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import './Register.css';
 import clearSelection from "../../functions/ClearSelection";
+import { Authenticate } from "../../functions/Auth";
 
 const Register = () => {
     const [email, setEmail] = useState("")
@@ -25,8 +26,6 @@ const Register = () => {
     }
 
     const onRegisterButtonClick = () => {
-
-        // Set initial error values to empty
         setEmailError("")
         setPasswordError("")
         setConfirmPasswordError("")
@@ -63,6 +62,7 @@ const Register = () => {
         }
 
         // register
+        Authenticate("http://localhost:8000/user/register/", email, password)
     }
 
     return <div className={"Register"}>
